:art: Move Etch components into separate folder
