+++ lib/about.js
@@ -70,7 +70,7 @@ export default class About {
 
   deserialize (state) {
     if (!this.views.aboutView) {
-      AboutView = AboutView || require('./about-view')
+      AboutView = AboutView || require('./components/about-view')
 
       this.setState(state)
