+++ lib/components/about-status-bar.js
@@ -4,7 +4,7 @@
 
 import {CompositeDisposable} from 'atom'
 import etch from 'etch'
-import EtchComponent from './etch-component'
+import EtchComponent from '../etch-component'
 
 export default class AboutStatusBar extends EtchComponent {
   constructor () {
