+++ lib/components/about-view.js
@@ -6,7 +6,7 @@ import {Disposable} from 'atom'
 import etch from 'etch'
 import shell from 'shell'
 import AtomLogo from './atom-logo'
-import EtchComponent from './etch-component'
+import EtchComponent from '../etch-component'
 import UpdateView from './update-view'
 
 export default class AboutView extends EtchComponent {
