+++ lib/components/atom-logo.js
@@ -3,7 +3,7 @@
 /* eslint-disable react/no-unknown-property */
 
 import etch from 'etch'
-import EtchComponent from './etch-component'
+import EtchComponent from '../etch-component'
 
 export default class AtomLogo extends EtchComponent {
   render () {
