+++ lib/components/update-view.js
@@ -3,8 +3,8 @@
 /* eslint-disable react/no-unknown-property */
 
 import etch from 'etch'
-import EtchComponent from './etch-component'
-import UpdateManager from './update-manager'
+import EtchComponent from '../etch-component'
+import UpdateManager from '../update-manager'
 
 export default class UpdateView extends EtchComponent {
   constructor (props) {
