+++ lib/main.js
@@ -61,7 +61,7 @@ export default {
 
   showStatusBarIfNeeded () {
     if (this.isUpdateAvailable() && this.statusBar) {
-      StatusBarView = StatusBarView || require('./about-status-bar')
+      StatusBarView = StatusBarView || require('./components/about-status-bar')
 
       let statusBarView = new StatusBarView()
